+++ src/components/ShadowToggle.tsx
@@ -1,6 +1,6 @@
 import React, { memo } from 'react';
 import { Pressable, View, type LayoutChangeEvent } from 'react-native';
-import { Canvas, RoundedRect, Shadow } from '@shopify/react-native-skia';
+import { Canvas, Shadow } from '@shopify/react-native-skia';
 import Animated, {
   interpolateColor,
   useAnimatedReaction,
@@ -27,6 +27,7 @@ import {
 } from '../utils';
 import LinearGradientFill from './ShadowLinearGradientFill';
 import { useAnimatedOffset } from '../hooks/useAnimatedOffset';
+import { CornerRadii } from './CornerRadii';
 
 const PressButton = Animated.createAnimatedComponent(Pressable);
 
@@ -64,7 +65,6 @@ export const UnifiedShadowToggle = memo(function ShadowToggle({
     reflectedLightBlur,
   });
 
-  const boxRadius = numerify(style?.borderRadius, 12);
   const isLinear = isLinearProps(props);
 
   const styleWidth = width ?? numerify(style?.width, 0);
@@ -138,13 +138,11 @@ export const UnifiedShadowToggle = memo(function ShadowToggle({
             { width: layoutSize.width * 1.4, height: layoutSize.height * 1.4 },
           ]}
         >
-          <RoundedRect
-            x={0}
-            y={0}
+          <CornerRadii
             width={layoutSize.width}
             height={layoutSize.height}
-            r={boxRadius}
-            color={animatedBackgroundColor} // The background fill of the rect
+            style={style}
+            backgroundColor={animatedBackgroundColor}
           >
             {isLinear && (
               <LinearGradientFill
@@ -169,7 +167,7 @@ export const UnifiedShadowToggle = memo(function ShadowToggle({
                 inner
               />
             )}
-          </RoundedRect>
+          </CornerRadii>
         </Canvas>
       )}
       <PressButton {...props} style={[style, COMMON_STYLES.canvasWrapper]}>